Add parameter 'subsample' to elastic transform
The most expensive part so far is the spline interpolation of the
control point coordinates to the size of the output volume. With
subsample > 1, this will be carried out on a downsampled volume and
resulting coordinates are linearly interpolated for the final output.

This is much faster and, depending on your control point spacing, may
come with no noticable difference.